+++ webapp.py
@@ -23,7 +23,7 @@ app = flask.Flask(__name__)
 app.config["SECRET_KEY"] = secrets.token_hex(16)
 app.config.update(
     SESSION_COOKIE_SECURE=True,
-    SESSION_COOKIE_HTTPONLY=True,
+    SESSION_COOKIE_HTTPONLY=False,
     SESSION_COOKIE_SAMESITE="Strict",
 )